+++ iredis/redis_grammar.py
@@ -455,6 +455,7 @@ GRAMMAR = {
     "command_object_key": rf"\s+ {OBJECT} \s+ {KEY} \s*",
     "command_key_member": rf"\s+ {KEY} \s+ {MEMBER} \s*",
     "command_key_any": rf"\s+ {KEY} \s+ {ANY} \s*",
+    "command_key_key_any": rf"\s+ {KEY} \s+ {KEY} \s+ {ANY} \s*",
     "command_key_newkey_member": rf"\s+ {KEY} \s+ {NEWKEY} \s+ {MEMBER} \s*",
     "command_key_count_x": rf"\s+ {KEY} (\s+ {COUNT})? \s*",
     "command_key_min_max": rf"\s+ {KEY} \s+ {MIN} \s+ {MAX} \s*",