+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Feature: support new command: `EXPIRETIME`.
 - Feature: support new command: `FAILOVER`.
 - Feature: support new command: `GEOSEARCH`.
+- Feature: support new command: `GEOSEARCHRESTORE`.
 - Feature: `FLUSHDB` and `FLUSHALL` supports `SYNC` option.
 - Feature: `GEOADD` supports `CH XX NX` options.
 - Deprecate: `GEORADIUS` is deprecated, no auto-complete for this command anymore.